requirements.txt
setuptools
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib==1.0.4
jiter==0.5.0
openai==1.40.8
pathlib==1.0.1
playsound==1.3.0
PyAudio==0.2.14
pydantic==2.8.2
pydantic_core==2.20.1
pyobjc-framework-Cocoa==10.3.1
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
SpeechRecognition==3.10.4
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2